# openapi.yaml
openapi: 3.0.3
info:
  title: Tereka Online API
  version: "1.0.0"
  description: >
    Multi-SACCO management platform API for Uganda. This is a curated specification of the
    integrator-facing surface (authentication, member self-service, mobile-money integration, and the
    core financial/loan read + approval endpoints). It is maintained by hand alongside the code; the
    authoritative behaviour is the Spring Boot backend under `backend-java`.


    ## Conventions

    - Base path: `/api/v1`.

    - Successful responses are wrapped in an envelope: `{ "data": <payload> }`.

    - Errors use `{ "error": { "timestamp", "status", "code", "message" } }`.

    - Authentication is a bearer token obtained from a login endpoint, sent as
      `Authorization: Bearer <token>`.

    - Abuse-prone endpoints are rate limited per client IP and may return `429` with a `Retry-After`
      header.
  contact:
    name: Tereka Online
servers:
  - url: "https://{host}/api/v1"
    description: Deployment
    variables:
      host:
        default: tereka.online
  - url: "http://127.0.0.1:8080/api/v1"
    description: Local backend
...
security:
  - bearerAuth: []
paths:
...
  /tenants:
    get:
      tags: [SACCO administration]
      summary: List SACCOs visible to the current staff user
      description: >
        Platform users see all registered SACCOs. SACCO staff see only their own SACCO.
      responses:
        "200":
          description: SACCO list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/TenantResponse" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
    post:
      tags: [SACCO administration]
      summary: Register a SACCO inside the platform
      description: >
        Platform administrators can create a SACCO directly. If `paymentStatus` is `paid`, the SACCO
        is activated immediately; otherwise activation waits for subscription payment.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTenantRequest" }
      responses:
        "201":
          description: SACCO created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/TenantResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /tenants/{tenantId}:
    get:
      tags: [SACCO administration]
      summary: Get one SACCO
      parameters:
        - { name: tenantId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: SACCO details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/TenantResponse" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /tenants/{tenantId}/status:
    patch:
      tags: [SACCO administration]
      summary: Update SACCO approval or operating status
      parameters:
        - { name: tenantId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTenantStatusRequest" }
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/TenantResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /tenants/{tenantId}/collection-mode:
    patch:
      tags: [SACCO administration]
      summary: Set the platform-approved collection mode for a SACCO
      description: >
        Platform Super Admin control for whether a SACCO may use no online collection, mobile-money
        collection only, bank-led collection only, or both. Disallowed channels are deactivated when
        the mode is changed.
      parameters:
        - { name: tenantId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCollectionModeRequest" }
      responses:
        "200":
          description: Collection mode updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/TenantResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /tenants/{tenantId}/collection-settings:
    patch:
      tags: [SACCO administration]
      summary: Activate or deactivate SACCO collection channels
      description: >
        SACCO administrators can activate mobile-money and/or bank collection only when the platform
        has allowed those channels through `/tenants/{tenantId}/collection-mode`.
      parameters:
        - { name: tenantId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCollectionSettingsRequest" }
      responses:
        "200":
          description: Collection settings updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/TenantResponse" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: Requested channel is not allowed by the platform
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Staff/platform session token from `/auth/login`.
    memberBearer:
      type: http
      scheme: bearer
      description: Member session token from `/member-auth/login`.
    callbackSignature:
      type: apiKey
      in: header
      name: X-Mobile-Money-Signature
      description: HMAC-SHA256 of the raw request body using the shared callback secret.
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            timestamp: { type: string, format: date-time }
            status: { type: integer, example: 400 }
            code: { type: string, example: "AUTH_INVALID" }
            message: { type: string }
...
    TenantResponse:
      type: object
      properties:
        id: { type: string, example: "tenant_green_valley" }
        name: { type: string, example: "Green Valley SACCO" }
        abbreviation: { type: string, example: "GVS" }
        registrationNo: { type: string, example: "REG-001" }
        district: { type: string, example: "Wakiso" }
        country: { type: string, example: "Uganda" }
        localeCode: { type: string, example: "en-UG" }
        currencyCode: { type: string, example: "UGX" }
        currencyDigits: { type: integer, example: 0 }
        licenseExpiry: { type: string, format: date }
        packageId: { type: string, example: "starter" }
        status:
          type: string
          enum: [pending_review, approved, active, suspended, terminated]
        allowedCollectionMode:
          $ref: "#/components/schemas/CollectionMode"
        mobileMoneyCollectionActive: { type: boolean }
        bankCollectionActive: { type: boolean }
        mobileMoneyCollectionAvailable:
          type: boolean
          description: True only when mobile money is both platform-allowed and SACCO-activated.
        bankCollectionAvailable:
          type: boolean
          description: True only when bank collection is both platform-allowed and SACCO-activated.
    CreateTenantRequest:
      type: object
      required: [name, abbreviation, licenseExpiry]
      properties:
        name: { type: string, example: "Green Valley SACCO" }
        abbreviation:
          type: string
          description: SACCO code used at login.
          example: "GVS"
        registrationNo: { type: string, example: "REG-001" }
        district: { type: string, example: "Wakiso" }
        parish: { type: string, example: "Nansana" }
        village: { type: string, example: "Central" }
        contactNumber: { type: string, example: "+256700123456" }
        memberRange:
          type: string
          example: "100-250"
        country: { type: string, example: "Uganda" }
        localeCode: { type: string, example: "en-UG" }
        currencyCode: { type: string, example: "UGX" }
        currencyDigits: { type: integer, example: 0 }
        licenseExpiry: { type: string, format: date }
        packageId: { type: string, example: "starter" }
        paymentStatus:
          type: string
          enum: [paid, pending]
          description: Paid activates the SACCO immediately; pending keeps it awaiting subscription payment.
    UpdateTenantStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending_review, approved, active, suspended, terminated]
    CollectionMode:
      type: string
      enum: [NONE, MOBILE_MONEY_ONLY, BANK_ONLY, BOTH]
      description: Platform-approved collection channels for a SACCO.
    UpdateCollectionModeRequest:
      type: object
      required: [allowedCollectionMode]
      properties:
        allowedCollectionMode:
          $ref: "#/components/schemas/CollectionMode"
    UpdateCollectionSettingsRequest:
      type: object
      properties:
        mobileMoneyActive:
          type: boolean
          description: SACCO requests mobile-money collection activation.
        bankActive:
          type: boolean
          description: SACCO requests bank-led collection activation.
...